Add weewx availability topic and last will
This adds a new option to publish the status of weewx and the MQTT
client. If enabled, the service will publish a new topic under
`topic/availability` with the value of "online" when the service is
connected and publishing messages. If the client disconnects the message
"offline" will be posted to indicate the other values are now stale.
This also takes advantage of the last will feature and will publish
"offline" if this service disconnects uncleanly.